+++ changelog.md
@@ -2,6 +2,7 @@
 
 ### [Unreleased]
 - Ensures Action Scheduler is loaded properly.
+- Adds additional schedule CLI command.
 
 ### [0.2.7] - 2025-11-30
 - Addresses minor feedback from Plugin Check.